#39 After Code Review: loadClass method extracted + RuntimeException changed to IllegalStateException
