+++ subzero-core/src/main/java/info/jerrinot/subzero/internal/ClassLoaderUtils.java
@@ -9,6 +9,7 @@ import java.lang.reflect.Method;
 
 public class ClassLoaderUtils {
     private static final boolean DEBUG_CLASSLOADING = Boolean.getBoolean("subzero.debug.classloading");
+    private static final Method loadClassMethodRef = getClassLoadMethod();
 
     public static ClassLoader getConfiguredClassLoader(HazelcastInstance hz) {
         try {
@@ -38,6 +39,42 @@ public class ClassLoaderUtils {
             } catch (InvocationTargetException e1) {
                 throw new IllegalArgumentException("Unknown instance object " + hz, e1);
             }
+        }
+    }
+
+    public static Class<?> loadClass(String className, ClassLoader classLoader) throws InvocationTargetException, IllegalAccessException {
+        return (Class<?>) loadClassMethodRef.invoke(null, classLoader, className);
+    }
+
+    private static Method getClassLoadMethod() {
+        Class<?> classLoaderUtilClass = getClassLoaderUtilClass();
+        try {
+            return classLoaderUtilClass.getDeclaredMethod("loadClass", ClassLoader.class, String.class);
+        } catch (NoSuchMethodException e) {
+            throw new IllegalStateException(e);
+        }
+    }
+
+    private static Class<?> getClassLoaderUtilClass() {
+        String[] possibleNames = {
+                "com.hazelcast.nio.ClassLoaderUtil",
+                "com.hazelcast.internal.nio.ClassLoaderUtil"
+        };
+
+        for (String possibleName : possibleNames) {
+            Class<?> loadedClass = loadClass(possibleName);
+            if (loadedClass != null) return loadedClass;
+        }
+        throw new IllegalStateException("unable to load ClassLoaderUtil");
+    }
+
+    private static Class<?> loadClass(String name) {
+        try {
+            return Class.forName(name);
+        } catch (ClassNotFoundException e) {
+            return null;
+        } catch (NoClassDefFoundError e) {
+            return null;
         }
     }
 }
