+++ subzero-core/src/main/java/info/jerrinot/subzero/internal/strategy/DelegatingClassResolver.java
@@ -1,9 +1,9 @@
 package info.jerrinot.subzero.internal.strategy;
 
 import com.esotericsoftware.kryo.util.DefaultClassResolver;
+import info.jerrinot.subzero.internal.ClassLoaderUtils;
 
 import java.lang.reflect.InvocationTargetException;
-import java.lang.reflect.Method;
 
 /**
  * Delegates class loading to Hazelcast -> SubZero will use the same strategy
@@ -13,7 +13,6 @@ import java.lang.reflect.Method;
 final class DelegatingClassResolver extends DefaultClassResolver {
 
     private final ClassLoader classLoader;
-    private static final Method loadClassMethodRef = getClassLoadMethod();
 
     public DelegatingClassResolver(ClassLoader classLoader) {
         this.classLoader = classLoader;
@@ -22,46 +21,15 @@ final class DelegatingClassResolver extends DefaultClassResolver {
     @Override
     protected Class<?> getTypeByName(String className) {
         try {
-            return (Class<?>) loadClassMethodRef.invoke(null, classLoader, className);
+            return ClassLoaderUtils.loadClass(className, classLoader);
         } catch (InvocationTargetException e) {
             if (e.getCause() instanceof ClassNotFoundException) {
                 return null;
             }
-            throw new RuntimeException(e);
+            throw new IllegalStateException(e);
         } catch (IllegalAccessException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private static Method getClassLoadMethod() {
-        Class<?> classLoaderUtilClass = getClassLoaderUtilClass();
-        try {
-            return classLoaderUtilClass.getDeclaredMethod("loadClass", ClassLoader.class, String.class);
-        } catch (NoSuchMethodException e) {
             throw new IllegalStateException(e);
         }
     }
 
-    private static Class<?> getClassLoaderUtilClass() {
-        String[] possibleNames = {
-                "com.hazelcast.nio.ClassLoaderUtil",
-                "com.hazelcast.internal.nio.ClassLoaderUtil"
-        };
-
-        for (String possibleName : possibleNames) {
-            Class<?> loadedClass = loadClass(possibleName);
-            if (loadedClass != null) return loadedClass;
-        }
-        throw new IllegalStateException("unable to load ClassLoaderUtil");
-    }
-
-    private static Class<?> loadClass(String name) {
-        try {
-            return Class.forName(name);
-        } catch (ClassNotFoundException e) {
-            return null;
-        } catch (NoClassDefFoundError e) {
-            return null;
-        }
-    }
 }
